feat(entry files): allows to redefine sets of the specs for distribution between browsers
- created utils.js and moved there isSpecFile function
- extracted splitArray to utils.js
- allowed to customize getSets function in sharding config
- added to README.md example for getSets customization
