+++ lib/framework.js
@@ -1,6 +1,4 @@
-const _ = require('lodash');
-
-const { isSpecFile } = require('./matchers');
+const { isSpecFile, splitArray } = require('./utils');
 
 function getConfig(fullConfig) {
   // ensure we can manipulate config settings
@@ -10,6 +8,7 @@ function getConfig(fullConfig) {
   config.base = config.base ? config.base : '/base';
   config.sets = [];
   config.indexMap = {};
+  config.getSets = config.getSets ? config.getSets : getSets;
   return config;
 }
 
@@ -27,23 +26,17 @@ function setupCoverageReporting(fullConfig) {
   fullConfig.coverageReporter.browserId = 'name';
 }
 
-// modified from example at:
-// https://stackoverflow.com/questions/8188548/splitting-a-js-array-into-n-arrays
-function splitArray(flatArray, numCols){
-  const maxColLength = Math.ceil(flatArray.length/numCols);
-  const nestedArray = _.chunk(flatArray, maxColLength);
-  for (var i = nestedArray.length; i < numCols; i++) {
-    nestedArray.push([]);
-  }
-  return nestedArray;
-}
-
 function setupSets(config, basePath, files) {
-  var specs = files.served.map(f => config.base + f.path.replace(basePath, ''))
-    .filter(p => isSpecFile(p, config.specMatcher));
-  config.sets = splitArray(specs, config.browserCount);
+  config.sets = config.getSets(config, basePath, files);
   var lastSet = config.sets && config.sets.length && config.sets[config.sets.length - 1];
   config.performSharding = (lastSet && lastSet.length);
+}
+
+// can be redefined in sharding config
+function getSets(config, basePath, files) {
+  var specs = files.served.map(f => config.base + f.path.replace(basePath, ''))
+    .filter(p => isSpecFile(p, config.specMatcher));
+  return splitArray(specs, config.browserCount);
 }
 
 function setupSharding(config, fullConfig, log) {
