+++ lib/middleware.js
@@ -1,4 +1,4 @@
-const { isSpecFile } = require('./matchers');
+const { isSpecFile } = require('./utils');
 
 const idParamExtractor = /\/\?id=(\d+)/;
 const idCookieExtractor = /browser.id=(\d+)/;
